Enemy now uses buffer of states

+++ src/net/mod.rs
@@ -4,7 +4,7 @@ use crate::input::InputState;
 
 pub const TICKRATE: u8 = 30;
 const TICKLEN_S: f32 = 1. / TICKRATE as f32;
-pub const DELAY: usize = 2;
+pub const DELAY: u16 = 2;
 pub const BUFFER_SIZE: usize = 32;  // matters for ACK, untested, must be pwr of 2
 
 #[derive(Resource)]
